deduplication rework
